Add support for expansion of @Param lists
The existing support for expanders in method parameters is
limited to converting a single value. The change applies the
expander individually to each item in a collection or array,
thus making it useful for multi-valued query parameters, for
instance. The old behaviour is preserved because no existing
expanders would have been converting collections to strings
(probably).