Add support for expansion of @Param lists (#403)
* Add support for expansion of @Param lists

The existing support for expanders in method parameters is
limited to converting a single value. The change applies the
expander individually to each item in a collection or array,
thus making it useful for multi-valued query parameters, for
instance. The old behaviour is preserved because no existing
expanders would have been converting collections to strings
(probably).

* Change Collection to Iterable and add tests

* Add test for null conversion to empty string as well

* Remove array handling and null handling