Supports context path when using Ribbon LoadBalancingTarget
Before, `LoadBalancingTarget` stripped out the path and only used
`URI.getScheme()` and `URI.getHost()` to generate the `Request`.

Now, it will add `URI.getPath()` to the `Request` as well; this is
useful if you want to interact with endpoints with a context-path.

Update README.md Travis CI to point to correct git repository