Change to configure JAXBContextFactory with a list of binding classes.
We should be able to to just provide a list of POJOs with JAXB annotations and not have
to provide an ObjectFactory or jaxb.index (which requires ObjectFafctory) or special JAXB
annotated package-info.java or schema.

The use-case is a project with only uses a few JAXB classes and we don't want the whole
JAXB package configuration, especially since JAXBContext.newInstance(...) supports a list
of binding classes.