Fixes Map Based Parameter Checking for Generic Subclasses (#689)
Fixes #665

When verifying that any of th `@*Map` annotations are in fact
`Map` instances, we were assumping that all values are direct
extension of a `Map` with generic type information intact.  When
using frameworks like Spring, it is possible to have `Map` objects
that do not expose type information, like `HttpHeaders`, which
directly extend from a `Map` with the type information static.

This added additional checking to the `checkMapKeys` function
to accomodate for `Map` subclasses without type information.  If
the map key information cannot be validated, we simply pass it
through.