Allow unresolved expressions in Query Parameter Name and Body Template
Expressions in Query Parameter names and in a Body Template will
now no longer be removed if they are not resolved.  For Query Template,
this change will prevent invalid query name/value pairs from being
generated.

For the Body Template, the documentation states that
unresolved should be preserved, yet the code did not match.