Fixes NPE when a Response does not provide headers (#855)
Fixes #853

There are some scenarios reported where a server does
not provide headers with the response.  While this is
not typically expected, it's simple enough for Feign to
be resilient to it.  This change checks the headers provided
in the builder and if none are provided, an empty map is
used in it's place.