Refactored RequestTemplate Body Handling
Fixes #857

Changes introduced in #821 deprecated a number of methods on `RequestTemplate`
whose usage throughout the rest of the library was quite prevalent.
While attempting to refactor that usage, a number of additional issues
related to #821 came to light.  The most important one is that the charset
at the `RequestTemplate` level applied only to the resulting `Request` and
did not need to be applied to the `Template` based classes.

It was this mistake, propagating the `RequestTemplate#charset` to the
`Template` objects, was the root cause of the NPE addressed in #821.

This change reverts #821 in favor of a simpler approach that addresses
the original issue and maintains the `Request.Body` abstraction, while
removing the `@deprecation` notices.

Revert "NPE when resolving a template with binary body (#821)"
This reverts commit 4da6d5cad674e4eeae5b56035ac1f94992499449.