Refactored Header and Query parameter JAXRS Contract Parsing (#896)
As part of 10.x, the `headers()` and `queries()` collections on the
`RequestTemplate` were made read only.  The `JAXRSContract` was still
attempting to manipulate those directly.  THis was missed due to a
use case not accounted for in the tests.  I've added the appropriate
use case and corrected the usage of the template.