No Longer prepend uri with slash if it is a query string (#889)
* No Longer prepend uri with slash if it is a query string

Fixes #887

Changes to target and uri parsing did not take into account
Empty Targets or URI provided methods.  In these scenarios,
the target contains the entire path and the uri will be the
query string only.  This change takes that into account and
no longer prepends a slash in these cases.

* Removed unnecessary import for Spring from Target Test