Replaced comma with Constant Delimiter in Template (#930)
Replaced comma with Constant Delimiter in Template

Fixes #924

Commas were used to identify iterable content, which conflicted when
a comma delimited literal was provided during expansion.  This change
switches commas for semi-colons, which are considered reserved secondary
delimiters in RFC 6750 and should not be used without being pct-encoded.

Should be a safer choice.