Migrating to Sonatype OSSRH from JCenter
Fixes #697

This change replaces JCenter with Sonatype OSSRH for snapshot and
release management.  To complete this change, we will need to add
additional information to Travis, specifically the GPG information
to sign the artifacts during a release.

It's my opinion that those values should be in Travis directly and
not here in this repository.  I'm open to other suggestions and
feedback on this approach.

Additional changes include:

* Updating BOM template to include SCM per OSSRH rules.
* Removed Bintray and GitHub credentials, they are no longer needed.
* Adding GPG Signing Values
* Code Signing Paths