Clarify availability of a Response Body in FeignException
Fixes #920

FeignException may contain the data from the response if the response is available and contains data.  However, the method `content` is ambiguious and does not reveal it's intent.  User's have expressed confusion as to if it is for the Request or the Response.

This change adds a new method `responseBody` to addresse this.  Use of content is now `@deprecated`.