Add support for SOAP Envelope declarations to be inherited (#1141)
Fixes #1127

In certain situations the declarations on the SOAP envelope are not
inherited by JAXB when reading the documents.  This is particularly
troublesome when it is not possible to correct the XML at the source.

To support this a new `useFirstChild` option has been added to the
`SOAPDecoder` builder that will use `SOAPBody#getFirstChild()`
instead of `SOAPBody#extractContentAsDocument()`.  This will allow
users to supply a `package-info.java` to manage the element namespaces
explicitly and define what should occur if the namespace declarations
are missing.