Restore Request.Body accessibility for compatibility
Fixes #1163

Restores `Request.Body.encoded` static function and
`RequestTemplate.requestBody` to restore compatibility.
Additionally, mark it as deprecated to encourage usage of
`RequestTemplate.body`.

`Request.Body` is an internal abstraction that was made built
in an effort to better manage request body data.  It was made
public too early and will be reverted back to it's internal
state in future releases.