GH-1172: Store expanded Header Template values as Literals (#1203)
Fixes #1172

When a `HeaderTemplate` is expanded, the result is placed onto a new
`RequestTemplate` via it's `header` method.  This results in the expanded
result being turned back into a `HeaderTemplate`, simply out of
convenience.  This behavior, while fine more general use cases, is
problematic if the header value contains braces `{` `}`, as this will
cause the app to consider these values as new expressions.

At this point in the expansion process, there is no need to evaluate
the expanded values again, so this change allows a `Template` to be
created from an existing List of `TemplateChunks`, allowing for callers
to provide explicit `Literal` or `Expression` chunks directly into a
`Template`, by passing the template parsing algorithms.

`RequestTemplate#expand` has been updated to apply this logic for
`HeaderTemplate` values only.